first attempt to creating a embedding generator

+++ LargeGraph/GenerateEmbeddings/GenerateGNNEmbeddings_old.py
@@ -88,12 +88,15 @@ def create_embs_dic(graph, emb_dic, map):
 
 import json
 
-def write_emd_to_file(graph,emb_dic,  to_write_path=None):
+def write_emd_to_file(graph,emb_dic,  to_write_path=None, overwrite_prev_embeddings=False):
   if(to_write_path==None):
     to_write_path=dir_to_large
   map= read_int_to_node_mapping(dir_to_large)
   to_write= create_embs_dic(graph, emb_dic, map)
   file_name=  os.path.join(to_write_path,"emb.json")
-  f = open(file_name, "w")
+  if overwrite_prev_embeddings:
+    f = open(file_name, "w")
+  else:
+    f = open(file_name, "a")
   json.dump(to_write, f)
   f.close()
